changed the package name to core.base
