--- core-base/designsystem/src/androidMain/kotlin/template/corebase/designsystem/theme/CMPTheme.android.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright 2025 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
- */
-package template.corebase.designsystem.theme
-
-import android.os.Build
-import androidx.compose.material3.ColorScheme
-import androidx.compose.material3.dynamicDarkColorScheme
-import androidx.compose.material3.dynamicLightColorScheme
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalContext
-
-@Composable
-actual fun selectColorScheme(
-    isDarkMode: Boolean,
-    dynamicColor: Boolean,
-    colorScheme: ColorScheme,
-): ColorScheme {
-    return when {
-        dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
-            val context = LocalContext.current
-            if (isDarkMode) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
-        }
-
-        else -> colorScheme
-    }
-}
