+++ core-base/designsystem/src/commonMain/kotlin/template/core/base/designsystem/component/variant/BottomAppBarVariant.kt
@@ -1,0 +1,15 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
+ */
+package template.core.base.designsystem.component.variant
+
+enum class BottomAppBarVariant {
+    BOTTOM_WITH_ACTIONS,
+    BOTTOM_CUSTOM,
+}
