+++ core-base/designsystem/src/commonMain/kotlin/template/core/base/designsystem/component/variant/ProgressIndicatorVariant.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
+ */
+package template.core.base.designsystem.component.variant
+
+enum class ProgressIndicatorVariant {
+    INDETERMINATE_LINEAR,
+    INDETERMINATE_CIRCULAR,
+    DETERMINATE_LINEAR,
+    DETERMINATE_CIRCULAR,
+}
