+++ core-base/designsystem/src/commonMain/kotlin/template/core/base/designsystem/theme/DarkThemeConfig.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
+ */
+package template.core.base.designsystem.theme
+
+enum class DarkThemeConfig(val configName: String) {
+    FOLLOW_SYSTEM("Follow System"),
+    LIGHT("Light"),
+    DARK("Dark"),
+    ;
+
+    companion object {
+        fun fromValue(value: String): DarkThemeConfig {
+            return entries.find { it.configName.equals(value, ignoreCase = true) } ?: FOLLOW_SYSTEM
+        }
+    }
+}
