--- core-base/designsystem/src/commonMain/kotlin/template/corebase/designsystem/component/CMPAlertDialog.kt
@@ -1,72 +1,0 @@
-/*
- * Copyright 2025 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
- */
-package template.corebase.designsystem.component
-
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.AlertDialogDefaults
-import androidx.compose.material3.BasicAlertDialog
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.window.DialogProperties
-import template.corebase.designsystem.component.variant.AlertDialogVariant
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CMPAlertDialog(
-    onDismissRequest: () -> Unit,
-    confirmButton: @Composable () -> Unit,
-    modifier: Modifier = Modifier,
-    dismissButton: @Composable (() -> Unit)? = null,
-    icon: @Composable (() -> Unit)? = null,
-    title: @Composable (() -> Unit)? = null,
-    text: @Composable (() -> Unit)? = null,
-    shape: Shape = AlertDialogDefaults.shape,
-    containerColor: Color = AlertDialogDefaults.containerColor,
-    iconContentColor: Color = AlertDialogDefaults.iconContentColor,
-    titleContentColor: Color = AlertDialogDefaults.titleContentColor,
-    textContentColor: Color = AlertDialogDefaults.textContentColor,
-    tonalElevation: Dp = AlertDialogDefaults.TonalElevation,
-    properties: DialogProperties = DialogProperties(),
-    variant: AlertDialogVariant = AlertDialogVariant.CUSTOM,
-    basicContent: @Composable (() -> Unit)? = null,
-) {
-    when (variant) {
-        AlertDialogVariant.CUSTOM -> AlertDialog(
-            onDismissRequest = onDismissRequest,
-            confirmButton = confirmButton,
-            dismissButton = dismissButton,
-            icon = icon,
-            title = title,
-            text = text,
-            shape = shape,
-            containerColor = containerColor,
-            iconContentColor = iconContentColor,
-            titleContentColor = titleContentColor,
-            textContentColor = textContentColor,
-            tonalElevation = tonalElevation,
-            properties = properties,
-            modifier = modifier,
-        )
-
-        AlertDialogVariant.BASIC -> basicContent?.let { content ->
-            BasicAlertDialog(
-                onDismissRequest = onDismissRequest,
-                modifier = modifier,
-                properties = properties,
-            ) {
-                content()
-            }
-        }
-    }
-}
