--- core-base/designsystem/src/commonMain/kotlin/template/corebase/designsystem/component/CMPScaffold.kt
@@ -1,85 +1,0 @@
-/*
- * Copyright 2025 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
- */
-package template.corebase.designsystem.component
-
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FabPosition
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.ScaffoldDefaults
-import androidx.compose.material3.contentColorFor
-import androidx.compose.material3.pulltorefresh.PullToRefreshBox
-import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CMPScaffold(
-    modifier: Modifier = Modifier,
-    topBar: @Composable () -> Unit = {},
-    bottomBar: @Composable () -> Unit = {},
-    snackbarHost: @Composable () -> Unit = {},
-    floatingActionButton: @Composable () -> Unit = {},
-    floatingActionButtonPosition: FabPosition = FabPosition.End,
-    containerColor: Color = MaterialTheme.colorScheme.background,
-    contentColor: Color = contentColorFor(containerColor),
-    contentWindowInsets: WindowInsets = ScaffoldDefaults.contentWindowInsets,
-    rememberPullToRefreshStateData: PullToRefreshStateData? = null,
-    content: @Composable (PaddingValues) -> Unit,
-) {
-    Scaffold(
-        modifier = modifier,
-        topBar = topBar,
-        bottomBar = bottomBar,
-        snackbarHost = snackbarHost,
-        floatingActionButton = floatingActionButton,
-        floatingActionButtonPosition = floatingActionButtonPosition,
-        containerColor = containerColor,
-        contentColor = contentColor,
-        contentWindowInsets = contentWindowInsets,
-    ) { paddingValues ->
-        if (rememberPullToRefreshStateData != null) {
-            val pullToRefreshState = rememberPullToRefreshState()
-            PullToRefreshBox(
-                state = pullToRefreshState,
-                isRefreshing = rememberPullToRefreshStateData.isRefreshing,
-                onRefresh = rememberPullToRefreshStateData.onRefresh,
-                modifier = Modifier.padding(paddingValues),
-            ) {
-                content(PaddingValues(0.dp))
-            }
-        } else {
-            content(paddingValues)
-        }
-    }
-}
-
-data class PullToRefreshStateData(
-    val isRefreshing: Boolean,
-    val onRefresh: () -> Unit,
-)
-
-@Composable
-fun rememberPullToRefreshStateData(
-    isRefreshing: Boolean = false,
-    onRefresh: () -> Unit = { },
-) = remember(isRefreshing, onRefresh) {
-    PullToRefreshStateData(
-        isRefreshing = isRefreshing,
-        onRefresh = onRefresh,
-    )
-}
