--- core-base/designsystem/src/commonMain/kotlin/template/corebase/designsystem/component/variant/ButtonVariant.kt
@@ -1,18 +1,0 @@
-/*
- * Copyright 2025 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
- */
-package template.corebase.designsystem.component.variant
-
-enum class ButtonVariant {
-    FILLED,
-    TONAL,
-    ELEVATED,
-    OUTLINED,
-    TEXT,
-}
